add unit tests for Hz and Temp regex in tsv_read, changed data type to floats

+++ ferro/data.py
@@ -273,7 +273,7 @@ class SampleData:
 
 
 class HysteresisData(SampleData):
-    def __init__(self, freq=100, **kwargs):
+    def __init__(self, freq=100.0, **kwargs):
         """
         Inherits SampleData. See that class for info on thickness, area, 
         and temperature
@@ -306,7 +306,7 @@ class HysteresisData(SampleData):
         if type(self) == type(other):
             if (self.area == other.area and
                     self.thickness == other.thickness and
-                    np.array_equal(self.voltage,other.voltage) and
+                    np.array_equal(self.voltage, other.voltage) and
                     np.array_equal(self.polarization, other.polarization) and
                     np.array_equal(self.time, other.time) and
                     np.array_equal(self.current, other.current)):
@@ -346,18 +346,18 @@ class HysteresisData(SampleData):
 
         r = re.compile(".*(_| )(\d+)C.*")
         try:
-            self.temp = int(r.match(filename).group(2)) + 273
+            self.temp = float(r.match(filename).group(2)) + 273
         except AttributeError:
             try:
                 r = re.compile(".*(\d+)K.*")
-                self.temp = int(r.match(filename).group(2))
+                self.temp = float(r.match(filename).group(2))
             except AttributeError:
                 print("No temperature specified. Defaulting to 300K")
                 next
 
         r = re.compile(".*(_| )(\d+)Hz.*")
         try:
-            self.freq = r.match(filename).group(2)
+            self.freq = float(r.match(filename).group(2))
         except AttributeError:
             print("No frequency specified. Defaulting to 100Hz")
             next
